Renamings and added documentation
