+++ src/client/app/konvolute/konvolut/konvolut.component.html
@@ -36,7 +36,6 @@
                     </div>
                     <br/>
                     <rae-konvolut-grid [poems_in_grid]="poems"></rae-konvolut-grid>
-                    <button md-raised-button (click)="loadMore()">Load more</button>
                   </div>
                 </div>
               </div>
